requirements.txt
requests==2.31.0
discord.py==2.3.2
python-dotenv==1.0.0
pytz==2023.3
beautifulsoup4==4.12.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-watch==4.2.0
flake8==6.1.0
black==23.9.1
mypy==1.6.1

